+++ neo/Implementations/Blockchains/LevelDB/LevelDBBlockchain.py
@@ -813,8 +813,8 @@ class LevelDBBlockchain(Blockchain):
 
             self.TXProcessed += len(block.Transactions)
 
-            for event in to_dispatch:
-                events.emit(event.event_type, event)
+        for event in to_dispatch:
+            events.emit(event.event_type, event)
 
     def PersistBlocks(self):