+++ neo/Prompt/Commands/tests/test_claim_command.py
@@ -97,6 +97,14 @@ class UserWalletTestCase(WalletFixtureTestCase):
 
         wallet = self.GetWallet3()
 
+        claim = ClaimGas(wallet, require_password=False, args=['1'])
+
+        self.assertTrue(claim)
+
+    def test_5_wallet_claim_ok(self):
+
+        wallet = self.GetWallet3()
+
         claim = ClaimGas(wallet, require_password=False)
 
         self.assertTrue(claim)