+++ neo/SmartContract/StateMachine.py
@@ -65,6 +65,8 @@ class StateMachine(StateReader):
         # commit storages right away
         if success:
             self.Commit()
+        else:
+            self.ResetState()
 
         super(StateMachine, self).ExecutionCompleted(engine, success, error)
 
@@ -75,6 +77,13 @@ class StateMachine(StateReader):
             self._assets.Commit(self._wb, False)
             self._contracts.Commit(self._wb, False)
             self._storages.Commit(self._wb, False)
+
+    def ResetState(self):
+        self._accounts.Reset()
+        self._validators.Reset()
+        self._assets.Reset()
+        self._contracts.Reset()
+        self._storages.Reset()
 
     def TestCommit(self):
         if self._storages.DebugStorage:
@@ -530,13 +539,14 @@ class StateMachine(StateReader):
 
         storage_key = StorageKey(script_hash=context.ScriptHash, key=key)
 
+        keystr = key
         if len(key) == 20:
             keystr = Crypto.ToAddress(UInt160(data=key))
 
-            self.events_to_dispatch.append(SmartContractEvent(SmartContractEvent.STORAGE_DELETE, [keystr],
-                                                              context.ScriptHash, Blockchain.Default().Height + 1,
-                                                              engine.ScriptContainer.Hash if engine.ScriptContainer else None,
-                                                              test_mode=engine.testMode))
+        self.events_to_dispatch.append(SmartContractEvent(SmartContractEvent.STORAGE_DELETE, [keystr],
+                                                          context.ScriptHash, Blockchain.Default().Height + 1,
+                                                          engine.ScriptContainer.Hash if engine.ScriptContainer else None,
+                                                          test_mode=engine.testMode))
 
         self._storages.Remove(storage_key.ToArray())