+++ neo/bin/api_server.py
@@ -45,7 +45,8 @@ from logzero import logger
 from twisted.logger import STDLibLogObserver, globalLogPublisher
 
 # Twisted and Klein methods and modules
-from twisted.internet import reactor, task
+from twisted.internet import reactor, task, endpoints
+from twisted.web.server import Site
 
 # neo methods and modules
 from neo.Core.Blockchain import Blockchain
@@ -137,7 +138,11 @@ def main():
         parser.print_help()
         return
 
-    # Setup depending on command line arguments. By default, the testnet settings are already loaded.
+    # Setting the datadir must come before setting the network, else the wrong path is checked at net setup.
+    if args.datadir:
+        settings.set_data_dir(args.datadir)
+
+    # Network configuration depending on command line arguments. By default, the testnet settings are already loaded.
     if args.config:
         settings.setup(args.config)
     elif args.mainnet:
@@ -149,8 +154,6 @@ def main():
     elif args.coznet:
         settings.setup_coznet()
 
-    if args.datadir:
-        settings.set_data_dir(args.datadir)
     if args.maxpeers:
         settings.set_max_peers(args.maxpeers)
 
@@ -207,17 +210,17 @@ def main():
     if args.port_rpc:
         logger.info("Starting json-rpc api server on http://%s:%s" % (args.host, args.port_rpc))
         api_server_rpc = JsonRpcApi(args.port_rpc)
-#        endpoint_rpc = "tcp:port={0}:interface={1}".format(args.port_rpc, args.host)
-#        endpoints.serverFromString(reactor, endpoint_rpc).listen(Site(api_server_rpc.app.resource()))
+        endpoint_rpc = "tcp:port={0}:interface={1}".format(args.port_rpc, args.host)
+        endpoints.serverFromString(reactor, endpoint_rpc).listen(Site(api_server_rpc.app.resource()))
 #        reactor.listenTCP(int(args.port_rpc), server.Site(api_server_rpc))
-        api_server_rpc.app.run(args.host, args.port_rpc)
+#        api_server_rpc.app.run(args.host, args.port_rpc)
 
     if args.port_rest:
         logger.info("Starting REST api server on http://%s:%s" % (args.host, args.port_rest))
         api_server_rest = RestApi()
-#        endpoint_rest = "tcp:port={0}:interface={1}".format(args.port_rest, args.host)
-#        endpoints.serverFromString(reactor, endpoint_rest).listen(Site(api_server_rest.app.resource()))
-        api_server_rest.app.run(args.host, args.port_rest)
+        endpoint_rest = "tcp:port={0}:interface={1}".format(args.port_rest, args.host)
+        endpoints.serverFromString(reactor, endpoint_rest).listen(Site(api_server_rest.app.resource()))
+#        api_server_rest.app.run(args.host, args.port_rest)
 
     reactor.run()