+++ CHANGELOG.rst
@@ -3,14 +3,30 @@ Changelog
 
 All notable changes to this project are documented in this file.
 
-[0.7.0] in progress
+[0.7.0-dev] in progress
 -----------------------
+- fix a bug with smart-contract parameter string parsing `#412 <https://github.com/CityOfZion/neo-python/issues/412>`_
 - fix ``StateMachine.Contract_Migrate`` and add tests
 - add ability to attach tx attrs to build command and testinvoke.  altered tx attr parsing
+- updated the install instructions present on ``docs``
+- fix issues with some JSON-RPC properties `#418 <https://github.com/CityOfZion/neo-python/issues/418>`_
+- added support for optionally chunking through GAS claims in prompt `#419 <https://github.com/CityOfZion/neo-python/issues/419>`_
+- support RPC and REST endpoints in parallel `#420 <https://github.com/CityOfZion/neo-python/issues/420>`_
+- Added new command ``tkn_history`` to the prompt. It shows the recorded history of transfers of a given NEP5 token, that are related to the open wallet.
+- fix current block lookup during smart contract event processing `#426 <https://github.com/CityOfZion/neo-python/issues/426>`_
+- fixed custom datadir setup for prompt and api-server
+- added ``mint`` smart-contract event to NotificationDB `#433 <https://github.com/CityOfZion/neo-python/pull/433>`_
+- Allow ``SmartContract`` to use a hash with or without 0x prefix
+- update to neo-boa v0.4.4
+- Update to `neo-python-core <https://github.com/CityOfZion/neo-python-core/blob/master/HISTORY.rst>`_ v0.4.8:
+
+  - Create wallets with ``np-utils --create-wallet``
+  - ``BigInteger(0)`` now is ``b'\x00'``
+
 
 
 [0.6.9] 2018-04-30
------------------------
+------------------
 - alter logging
 - fix issue with dispatching transfer events when ``from_addr`` is ``False``
 - add TPS monitor to ``prompt`` ``state`` command
@@ -25,7 +41,7 @@ All notable changes to this project are documented in this file.
 
 
 [0.6.8] 2018-04-26
------------------------
+------------------
 - add ``ServiceEnabled`` boolean to settings to determine whether nodes should send other nodes blocks
 - updated new block retrieval mechanism
 - fix for token_delete command not removing tokens from wallet file
@@ -41,7 +57,7 @@ All notable changes to this project are documented in this file.
 
 
 [0.6.7] 2018-04-06
------------------------
+------------------
 - Update all the requirements
 - Networking changes
 - added ``--maxpeers`` option for ``np-prompt`` and ``np-api-server``.  This allows p2p discovery of new nodes up to the value specified